added wdt config func

+++ msp430/firmware/types.h
@@ -2,13 +2,13 @@
 #define _TYPES_
 
 typedef unsigned char uint8_t;
-typedef unsigned short uint16_t;
+typedef unsigned int uint16_t;
 typedef unsigned long uint32_t;
 typedef unsigned long size_t;
 typedef unsigned long long uint64_t;
 
 typedef char int8_t;
-typedef short int16_t;
+typedef int int16_t;
 typedef long int32_t;
 typedef long long int64_t;
